+++ service-worker.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = "linuxplus-trainer-rev6-7";
+const CACHE_NAME = "linuxplus-trainer-rev6-8";
 const APP_FILES = [
   "./",
   "./index.html",
@@ -11,6 +11,7 @@ const APP_FILES = [
   "./js/rev65.js",
   "./js/rev66.js",
   "./js/rev67.js",
+  "./js/rev68.js",
   "./icons/apple-touch-icon.png",
   "./icons/icon-192.png",
   "./icons/icon-512.png",